requirements.txt
pyarrow
dash
dash-leaflet
geopandas